requirements.txt
google-generativeai
streamlit
pandas
numpy
fuzzywuzzy
python-Levenshtein
openai

